+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unreleased -- v0.10.4
 ### Added
 - Updated format/filter mappings in APIUtils class
+- Work title and author info to edition endpoint
 ### Fixed
 -